Added new fragment for groups a user has joined
Groups are now also added to the users tree when sending
a join request, rather than on approval. This was done in order for
users to view the groups they have joined, the next step would be to
only show chats for users who have been approved and display that the
group they have joined is awaiting approval from that group's admin.